feat: produce layout/inspector on desktop and unify modifier-token resolution
Closes #1903 (the #1897 consolidation follow-up).

- Desktop now produces `layout/inspector`. The desktop daemon registered
  the kind but never wrote `layout-inspector.json`, so `data/fetch` degraded
  to `NotAvailable`. The desktop RenderEngine now writes it from the same
  captured `ImageComposeScene` semantics root + slot tables it already uses
  for `compose/semantics`, via a new CMP-portable
  `LayoutInspectorDataProducer.writeArtifacts(root, slotTables, density)`
  overload. The Compose `LayoutNode` child-walk is hardened for the
  skiko/desktop build, whose internal accessors mangle to `$ui` (not
  `$ui_release`) and return a `MutableVector` (not `Iterable`) — without that
  the desktop tree was a lone root node.

- Modifier-token resolution is computed once. The modifier ->
  {backgroundColor, borderColor, cornerRadius, shape, gap, padding}
  projection moves to a single `ModifierTokenResolver`. Both
  `compose/semantics` and `layout/inspector` feed it per-node inputs instead
  of each re-walking `getModifierInfo()` with a private copy.

- Canonical home for design tokens decided. `layout/inspector` is the
  canonical surface (tokens are modifier-derived, and it already models the
  modifier chain); `compose/semantics` keeps mirroring the byte-identical
  object for the design-parity token-compliance consumer, so that contract is
  untouched. `LayoutInspectorNode` gains `tokens`; schemaVersion 1 -> 2
  (additive). The `layout*` text-styling fields stay on `compose/semantics`:
  they come from the `GetTextLayoutResult` semantics action, not modifiers.

- Docs: DATA-PRODUCTS.md matrix flips desktop `layout/inspector` to producer,
  documents the resolved home decision, and bumps the schema notes.

Test: data-layoutinspector-connector unit tests; new desktop
DesktopLayoutInspectorTest (non-empty tree + resolved tokens) and the
existing DesktopSemanticsTokensTest pass; ktfmtCheckAll clean. Android
:daemon:android compile/tests rely on CI (no Android SDK in this sandbox).
